+++ packages/yourloops/components/menus/drawer/main-drawer.tsx
@@ -45,12 +45,12 @@ import HistoryIcon from '@material-ui/icons/History'
 import Divider from '@material-ui/core/Divider'
 import Box from '@material-ui/core/Box'
 
-import MedicalServiceIcon from '../../icons/MedicalServiceIcon'
-import PendingIcon from '../../icons/PendingIcon'
-import { PatientFilterTypes } from '../../../models/generic'
+import MedicalServiceIcon from '../../icons/medical-service-icon'
+import PendingIcon from '../../icons/pending-icon'
 import DrawerLinkItem from './drawer-link-item'
 import useMainDrawer from './main-drawer.hook'
 import DrawerCategoryItem from './drawer-category-item'
+import { PatientFilterTypes } from '../../../lib/patient/models/enums/patient-filter-type.enum'
 
 export interface MainDrawerProps {
   miniVariant?: boolean