+++ packages/yourloops/lib/metrics.ts
@@ -28,7 +28,7 @@
 import _ from 'lodash'
 import bows from 'bows'
 
-import config from './config'
+import config from './config/config'
 import { User } from './auth'
 
 type VariableScope = 'page' | 'visit'