+++ CHANGELOG.md
@@ -26,13 +26,14 @@ It is based on Tidepool Blip 1.27.
 - YLP-1906-Fix-training-modal-CSS
 - YLP-1916 Migrate tooltips from viz to dumb
 - YLP-1935 CBG thresholds should be used from medical data service
+- YLP-1942 Rename files in yourloops and medical-domain
 - YLP-1945 Migrate NoBar component to dumb
+- YLP-1960 enhance pipeline
 - Automate copyright generation on new and updated files
 - Changing CI from Jenkins to GitHub Actions
 - Refactoring profile page code and adding tests
 - Remove react-test-renderer dependency
 - Update copyright on all files in `dumb` and `yourloops` packages
-- YLP-1960 enhance pipeline
 
 ## 3.0.1 - 2022-10-11
 ### Added